test(ks,backend): move email preview permissions-policy test
The new test queried EmailTemplate rows that are only seeded by
applications/tests/conftest.py. Run it under applications/tests so
it gets the same fixtures as the CSP email preview test.

refs: YJDH-778
